pyproject.toml
[project]
name = "roadmap"
authors = [
    {name = "Uroflavin", email = "urovlavin@gmail.com"},
]
description = "Manage your project and team road maps in YAML"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["roadmap", "json", "schema", "markdown", "graphviz"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "python-dotenv==1.0.0",
    "pyyaml==6.0",
    "Jinja2==3.1.2",
    "pymdown-extensions==10.5",
    "jinja_markdown==1.210911",
    "MarkupSafe>=2.0",
    "jsonschema~=4.20.0",
    "pre-commit==3.6.0",
]
dynamic = ["version"]

[project.scripts]
roadmap = "roadmap_app.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
